ml/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
scikit-learn==1.4.1
xgboost==2.0.3
shap==0.45.0
prophet==1.1.5
pandas==2.2.1
numpy==1.26.4
pydantic==2.6.4
httpx==0.27.0
python-dotenv==1.0.0
joblib==1.3.2
